pyproject.toml
[project]
name = "rl-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gymnasium>=1.0",
    "numpy>=1.26",
    "optuna>=4.0",
    "polars>=1.0",
    "dbutils>=3.1",
    "pymysql>=1.1",
    "python-dotenv>=1.0",
    "rich-click>=1.9.7",
    "torch>=2.10",
    "zstandard>=0.23",
    "sbx-rl>=0.21.0",
    "jax[cuda13]>=0.9.1",
    "tfp-nightly[jax]>=0.26.0.dev20260308",
    "flax>=0.12.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rl_model"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu130" }

[project.scripts]
rl-model = "rl_model.main:cli"
